+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Utility to generate LEM font and palette from image
+
 ## [0.2.0]
 
 ### Added